@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=7.0"
     },
     "autoload": {
         "psr-4": {"BlueM\\": "src/"}
